% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GO.db}
\alias{GO.db}
\title{self-describing object for GO interface}
\format{
An object of class \code{GOparq} of length 1.
}
\usage{
GO.db
}
\description{
self-describing object for GO interface
}
\keyword{datasets}
