% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPANCESTOR}
\alias{GOBPANCESTOR}
\title{self-describing object for GO ANCESTOR BP}
\format{
An object of class \code{environment} of length 24546.
}
\usage{
GOBPANCESTOR
}
\description{
self-describing object for GO ANCESTOR BP
}
\keyword{datasets}
