% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPCHILDREN}
\alias{GOBPCHILDREN}
\title{environment for GO CHILDREN BP}
\format{
An object of class \code{environment} of length 14658.
}
\usage{
GOBPCHILDREN
}
\description{
environment for GO CHILDREN BP
}
\keyword{datasets}
