% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPPARENTS}
\alias{GOBPPARENTS}
\title{environment for GO PARENTS BP}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOBPPARENTS
}
\description{
environment for GO PARENTS BP
}
\keyword{datasets}
