% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOCCANCESTOR}
\alias{GOCCANCESTOR}
\title{self-describing object for GO ANCESTOR CC}
\format{
An object of class \code{environment} of length 4068.
}
\usage{
GOCCANCESTOR
}
\description{
self-describing object for GO ANCESTOR CC
}
\keyword{datasets}
