% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOCCCHILDREN}
\alias{GOCCCHILDREN}
\title{environment for GO CHILDREN CC}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOCCCHILDREN
}
\description{
environment for GO CHILDREN CC
}
\note{
Previous versions of GO.db used 'isa' to name elements
of this environment instead of 'is_a' which is
now used in go.obo, and here.
}
\keyword{datasets}
