% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOCCOFFSPRING}
\alias{GOCCOFFSPRING}
\title{environment for GO OFFSPRING CC}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOCCOFFSPRING
}
\description{
environment for GO OFFSPRING CC
}
\keyword{datasets}
