% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOCCPARENTS}
\alias{GOCCPARENTS}
\title{environment for GO PARENTS CC}
\format{
An object of class \code{environment} of length 4057.
}
\usage{
GOCCPARENTS
}
\description{
environment for GO PARENTS CC
}
\keyword{datasets}
