% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOMFANCESTOR}
\alias{GOMFANCESTOR}
\title{self-describing object for GO ANCESTOR MF}
\format{
An object of class \code{environment} of length 10122.
}
\usage{
GOMFANCESTOR
}
\description{
self-describing object for GO ANCESTOR MF
}
\keyword{datasets}
