% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOMFCHILDREN}
\alias{GOMFCHILDREN}
\title{environment for GO CHILDREN MF}
\format{
An object of class \code{environment} of length 2928.
}
\usage{
GOMFCHILDREN
}
\description{
environment for GO CHILDREN MF
}
\note{
Previous versions of GO.db used 'isa' to name elements
of this environment instead of 'is_a' which is
now used in go.obo, and here.
}
\keyword{datasets}
