% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOMFOFFSPRING}
\alias{GOMFOFFSPRING}
\title{environment for GO OFFSPRING MF}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOMFOFFSPRING
}
\description{
environment for GO OFFSPRING MF
}
\keyword{datasets}
