% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOMFPARENTS}
\alias{GOMFPARENTS}
\title{self-describing object for GO PARENTS MF}
\format{
An object of class \code{environment} of length 10142.
}
\usage{
GOMFPARENTS
}
\description{
self-describing object for GO PARENTS MF
}
\keyword{datasets}
