% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOparq_S4.R
\docType{class}
\name{GOparq-class}
\alias{GOparq-class}
\title{class for parquet representation of GO}
\description{
class for parquet representation of GO
}
\note{
This class just serves as a bridge to parquet files that emulate
the former SQLite schema of AnnotationDbi-based GO.db.
}
