% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOSYNONYM}
\alias{GOSYNONYM}
\title{environment for GO SYNONYM}
\format{
An object of class \code{environment} of length 26114.
}
\usage{
GOSYNONYM
}
\description{
environment for GO SYNONYM
}
\note{
The behavior is unlike that of GO.db::GOSYNONYM which gives
stanzas from OBO.  Synonymous phrases are returned along with ids
of secondary GO terms.
}
\examples{
get("GO:0000215", GO.db3::GOSYNONYM)
# see  https://www.ebi.ac.uk/QuickGO/term/GO:0000215 for related information
}
\keyword{datasets}
