% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOTERM}
\alias{GOTERM}
\title{environment for GO TERM}
\format{
An object of class \code{GOparqMap} of length 1.

An object of class \code{GOparqMap} of length 1.
}
\usage{
GOTERM

GOTERM
}
\description{
environment for GO TERM

environment for GOTERM
}
\keyword{datasets}
