% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOparq_S4.R
\name{keys,GOparq-method}
\alias{keys,GOparq-method}
\title{list keys}
\usage{
\S4method{keys}{GOparq}(x, keytype, ...)
}
\arguments{
\item{x}{instance of GOparq}

\item{keytype}{character(1) one of "GOID", "TERM", "DEFINITION", "ONTOLOGY"}

\item{\dots}{not used}
}
\description{
list keys
}
\examples{
head(keys(GO.db))
}
