% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4code.R
\name{keys,GOparqMap-method}
\alias{keys,GOparqMap-method}
\title{keylist}
\usage{
\S4method{keys}{GOparqMap}(x, keytype, ...)
}
\arguments{
\item{x}{instance of GOparqMap}

\item{keytype}{character}

\item{\dots}{not used}
}
\description{
keylist
}
