% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIds.R
\name{mapIds,GOparq-method}
\alias{mapIds,GOparq-method}
\title{mapIds method}
\usage{
\S4method{mapIds}{GOparq}(x, keys, column, keytype, ..., multiVals)
}
\arguments{
\item{x}{instance of GOparq}

\item{keys}{character vector}

\item{column}{single character element}

\item{keytype}{single character element among value of keytypes}

\item{\dots}{not used}

\item{multiVals, }{only relevant for keytype="ONTOLOGY", can take values 'first' (default),
'list', or 'CharacterList'}
}
\description{
mapIds method
}
\examples{
mapIds(GO.db3::GO.db, keys=c("MF", "BP"), column="GOID", keytype="ONTOLOGY", multiVals="CharacterList")
}
