% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOparq_S4.R
\name{select,GOparq-method}
\alias{select,GOparq-method}
\title{emulate AnnotationDbi}
\usage{
\S4method{select}{GOparq}(x, keys, columns, keytype, ...)
}
\arguments{
\item{x}{instance of GOparq}

\item{keys}{for filtering, if NULL, no filtering performed}

\item{columns}{for joining and selecting}

\item{keytype}{for table selection: one of "term", "goid"; all caps may be used}

\item{\dots}{not used}
}
\value{
data.frame
}
\description{
emulate AnnotationDbi
}
\examples{
select(GO.db, keys="low-affinity zinc ion transmembrane transporter activity", keytype="term",
    columns=c("GOID", "DEFINITION"))
select(GO.db, keys="GO:0009435", keytype="GOID", columns="TERM")
select(GO.db, keys="GO:0009435", keytype="GOID", columns=c("GOID", "TERM", "ONTOLOGY"))
}
