% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select3.R
\name{select3}
\alias{select3}
\title{superficial emulator of AnnotationDbi::select for Gene Ontology}
\usage{
select3(x, keys, columns = NULL, keytype, ...)
}
\arguments{
\item{x}{character, must be "GO.db3"}

\item{keys}{for filtering, if NULL, no filtering performed}

\item{columns}{for joining and selecting}

\item{keytype}{for table selection: one of "term", "goid"; all caps may be used}
}
\value{
data.frame
}
\description{
superficial emulator of AnnotationDbi::select for Gene Ontology
}
\note{
GO.db permitted columns and keytypes "DEFINITION", "GOID", "ONTOLOGY", "TERM"
}
\examples{
select3("GO.db3", keys="low-affinity zinc ion transmembrane transporter activity", keytype="term")
select3("GO.db3", keys="GO:0009435", keytype="GOID") |> dplyr::select(-definition)
select3("GO.db3", keys="GO:0009435", keytype="goid", columns=c("GOID", "TERM", "ONTOLOGY"))
}
