% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4code.R
\name{[[,GOparqMap-method}
\alias{[[,GOparqMap-method}
\title{selector}
\usage{
\S4method{[[}{GOparqMap}(x, i, j, ...)
}
\arguments{
\item{x}{instance of GOparqMap}

\item{i}{character}

\item{j}{character, likely not used}

\item{\dots}{not used}
}
\description{
selector
}
