% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPANCESTOR}
\alias{GOBPANCESTOR}
\title{environment for GO ANCESTOR BP}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOBPANCESTOR
}
\description{
environment for GO ANCESTOR BP
}
\keyword{datasets}
