% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPCHILDREN}
\alias{GOBPCHILDREN}
\title{environment for GO CHILDREN BP}
\format{
An object of class \code{environment} of length 14658.
}
\usage{
GOBPCHILDREN
}
\description{
environment for GO CHILDREN BP
}
\note{
Previous versions of GO.db used 'isa' to name elements
of this environment instead of 'is_a' which is
now used in go.obo, and here.
}
\keyword{datasets}
