% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPOFFSPRING}
\alias{GOBPOFFSPRING}
\title{environment for GO OFFSPRING BP}
\format{
An object of class \code{GOparqMap} of length 1.
}
\usage{
GOBPOFFSPRING
}
\description{
environment for GO OFFSPRING BP
}
\keyword{datasets}
