% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOBPPARENTS}
\alias{GOBPPARENTS}
\title{self-describing object for GO PARENTS BP}
\format{
An object of class \code{environment} of length 25152.
}
\usage{
GOBPPARENTS
}
\description{
self-describing object for GO PARENTS BP
}
\keyword{datasets}
