% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs.R
\name{GOBPPARENTS}
\alias{GOBPPARENTS}
\title{emulate the GO.db environments, but as functions: GOBP}
\usage{
GOBPPARENTS()
}
\value{
environment mapping from GOID to parents in BP ontology
}
\description{
emulate the GO.db environments, but as functions: GOBP
}
