% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOCCCHILDREN}
\alias{GOCCCHILDREN}
\title{environment for GO CHILDREN CC}
\format{
An object of class \code{environment} of length 1358.
}
\usage{
GOCCCHILDREN
}
\description{
environment for GO CHILDREN CC
}
\keyword{datasets}
