% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOMFCHILDREN}
\alias{GOMFCHILDREN}
\title{environment for GO CHILDREN MF}
\format{
An object of class \code{environment} of length 2928.
}
\usage{
GOMFCHILDREN
}
\description{
environment for GO CHILDREN MF
}
\keyword{datasets}
