% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GOSYNONYM}
\alias{GOSYNONYM}
\title{environment for GO SYNONYM}
\format{
An object of class \code{environment} of length 3312.
}
\usage{
GOSYNONYM
}
\description{
environment for GO SYNONYM
}
\examples{
get("GO:1905121", GOSYNONYM)
}
\keyword{datasets}
