% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs.R
\name{GOSYNONYM}
\alias{GOSYNONYM}
\title{emulate the GO.db environments, but as functions: GOSYNONYM}
\usage{
GOSYNONYM()
}
\value{
environment mapping from GOID to synonyms
}
\description{
emulate the GO.db environments, but as functions: GOSYNONYM
}
\examples{
head(unname(GOSYNONYM()[["GO:0009435"]]))
# this could be better
tmpcon = arrow::open_dataset(system.file("extdata", "go323", package="GO.db3"))
syn = arrow::open_dataset(grep("synonym", tmpcon$files, value=TRUE))
syn |> dplyr::filter(go_id == "GO:0009435") |> dplyr::collect()
}
